Simplify validate-npm-name.ts file
