+++ src/helpers/get-project-name-of-path.ts
@@ -1,0 +1,5 @@
+import path from "path"
+
+export const getProjectNameOfPath = (projectPath: string): string => {
+  return path.basename(path.resolve(projectPath))
+}
