+++ src/questionnaire/questions/validate-project-path.ts
@@ -1,9 +1,9 @@
 import Command from "@oclif/command"
 import { throwError } from "../../error-handling"
-import { validateProjectPath } from "../../helpers/validate-project-name"
+import { getProjectNameOfPath } from "../../helpers/get-project-name-of-path"
+import { validateNpmName } from "../../helpers/validate-npm-name"
 
 /**
- *
  * @param this Current Command instance
  * @param projectPath The project path the user input
  * @returns `true` if valid. If invalid, an error message of type `string` explaining the invalidity.
@@ -17,7 +17,8 @@ export function validateProjectPathInput(
       throw new TypeError("Expected projectPath to be a string.")
     }
 
-    const validation = validateProjectPath(projectPath.trim())
+    const projectName = getProjectNameOfPath(projectPath)
+    const validation = validateNpmName(projectName)
 
     if (!validation.valid) {
       const firstProblem = validation.problems[0]
